backend/requirements.txt
python-dotenv==1.1.0
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
python-jose==3.3.0
aiomysql==0.2.0
passlib[bcrypt]==1.7.4
yamale==5.2.1
numpy==1.26.4
psycopg2-binary==2.9.10
alembic==1.15.2
asyncpg==0.30.0
